@@ -1,10 +1,8 @@
 module github.com/fyne-io/gl-js
 
-go 1.14
+go 1.19
 
 require (
-	github.com/go-gl/gl v0.0.0-20211210172815-726fda9656d6
-	github.com/gopherjs/gopherjs v0.0.0-20211219123610-ec9572f70e60
-	github.com/goxjs/gl v0.0.0-20210104184919-e3fafc6f8f2a // indirect
-	golang.org/x/mobile v0.0.0-20211207041440-4e6c2922fdee
+	github.com/go-gl/gl v0.0.0-20231021071112-07e5d0ea2e71
+	golang.org/x/mobile v0.0.0-20231127183840-76ac6878050a
 )
